add like to post corrected

+++ src/main/java/project/controllers/rest/ApiLikesController.java
@@ -9,12 +9,15 @@ import project.dto.responseDto.LikeUsersListDto;
 import project.dto.responseDto.ResponseDto;
 import project.handlerExceptions.BadRequestException400;
 import project.models.Person;
+import project.models.Post;
 import project.models.PostLike;
 import project.security.TokenProvider;
 import project.services.PostLikeService;
+import project.services.PostService;
 
 import javax.servlet.ServletRequest;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -22,6 +25,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ApiLikesController {
 
+    private PostService postService;
     private PostLikeService postLikeService;
     private TokenProvider tokenProvider;
 
@@ -45,17 +49,20 @@ public class ApiLikesController {
 
         Person person = tokenProvider.getPersonByRequest(servletRequest);
 
-        List<Integer> personsWhoLikedPost = postLikeService.getAllPersonIdWhoLikedPost(addLikeDto.getItemId());
+        List<Integer> personsWhoLikedItem = new ArrayList<>();
 
-        if (personsWhoLikedPost.contains(person.getId())) throw new BadRequestException400();
+        if (addLikeDto.getType().equals("Post")) {
+            postLikeService.addLike(person.getId(), addLikeDto.getItemId());
+            postService.setMyLike(addLikeDto.getItemId(), true);
 
-        postLikeService.addLike(person.getId(), addLikeDto.getItemId());
-
-        personsWhoLikedPost.add(person.getId());
+             personsWhoLikedItem.addAll(postLikeService.getAllPersonIdWhoLikedPost(addLikeDto.getItemId()));
+        } else {
+            personsWhoLikedItem.addAll(null);
+        }
 
         return ResponseEntity.ok(
                 new ResponseDto<>(
-                new LikeUsersListDto(personsWhoLikedPost.size(), personsWhoLikedPost)));
+                new LikeUsersListDto(personsWhoLikedItem.size(), personsWhoLikedItem)));
     }
 
     @Secured("ROLE_USER")
@@ -64,7 +71,14 @@ public class ApiLikesController {
 
         Person person = tokenProvider.getPersonByRequest((HttpServletRequest) servletRequest);
 
-        Integer likeCount = postLikeService.deleteLike(addLikeDto.getItemId(), person.getId());
+        Integer likeCount;
+
+        if (addLikeDto.getType().equals("Post")) {
+            postService.setMyLike(addLikeDto.getItemId(), false);
+            likeCount = postLikeService.deleteLike(addLikeDto.getItemId(), person.getId());
+        } else {
+            likeCount = 0;
+        }
 
         return ResponseEntity.ok(new ResponseDto<>(likeCount));
     }
